% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{benetton.associations}
\alias{benetton.associations}
\title{Benetton Associations}
\format{A \link{\code{data.frame}}.}
\source{
Adapted from Pinson, Christian, and Vikas Tibrewala (1996), “United Colors of Benetton,” INSEAD-CEDEP case study.
}
\usage{
benetton.associations
}
\description{
Binary associations for Benetton (n = 2,151).
}
\keyword{datasets}

