% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{benetton}
\alias{benetton}
\title{Benetton}
\format{
A \code{\link{data.frame}}.
}
\source{
Pinson, Christian, and Vikas Tibrewala (1996), "United Colors of Benetton," INSEAD-CEDEP case study.
}
\usage{
benetton
}
\description{
Sales and advertising data for Benetton.
}
\keyword{datasets}
