% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{big5cola}
\alias{big5cola}
\title{Big 5 Cola}
\format{A \link{\code{data.frame}}.}
\usage{
big5cola
}
\description{
A table showing mean Big 5 personality scores according to user's most frequently consumed brand. Computed from \link{cola.original}.
}
\keyword{datasets}

