% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{brand.associations}
\alias{brand.associations}
\title{Brand Associations}
\format{A \code{\link{data.frame}}.}
\source{
Numbers International Pty Ltd.
}
\usage{
brand.associations
}
\description{
Binary associations between brands and attributes, collected from a sample of 3173 Australians in 2005 (random allocaiton of brand to attributes).
}
\keyword{datasets}

