% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{breakfast.dissimilarities}
\alias{breakfast.dissimilarities}
\title{Breakfast dissimilarities}
\format{An object of class \code{list} of length 4.}
\usage{
breakfast.dissimilarities
}
\description{
@description breakfast dissimilarity dataset.
 @format A \code{\link{matrix}}
}
\keyword{datasets}
