% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{burger.brand.tracking}
\alias{burger.brand.tracking}
\title{Burger Brand Tracking}
\format{
A \code{\link{data.frame}}. This data set was used as an example of stacking data in
  regression and driver analysis. In particular, Q4 and Q14 as the stacked outcome and predictors
  respectively.
}
\source{
30 June 2019: https://wiki.q-researchsoftware.com/wiki/File:Burger_Brand_Tracking_Oct_2017.sav Added to flipExampleData 28th August 2020
}
\usage{
burger.brand.tracking
}
\description{
Survey data on customer visits and preferences for meals at different burger
  restaurants brands.
}
\keyword{datasets}
