% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{car.switching}
\alias{car.switching}
\title{Car Switching}
\format{
An object of class \code{data.frame} with 17 rows and 17 columns.
}
\usage{
car.switching
}
\description{
@description Switching between brands of car.
 @references Colombo, Richard, Andrew Ehrenberg, and Darius Sabavala (1993), "The Car Challenge: Diversity in Analyzing Brand Switching Tables," Working Paper
 @format A \code{\link{matrix}}
}
\keyword{datasets}
