% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cola.original}
\alias{cola.original}
\title{Cola Original}
\format{A \link{\code{data.frame}}.}
\usage{
cola.original

cola.original
}
\description{
A data file from a survey of the Australian cola market in 2007.

A data file from a survey of the Australian cola market in 2007. This is a modified version of \link{cola.original}.
}
\keyword{datasets}

