% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cola.perceptions.large}
\alias{cola.perceptions.large}
\title{Cola Perceptions Large}
\format{A \link{\code{data.frame}}.}
\usage{
cola.perceptions.large
}
\description{
A table showing the counts of people to choose each attribute for each brand in Q5 of \link{cola.original}.
}
\keyword{datasets}

