% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cola.perceptions}
\alias{cola.perceptions}
\title{Cola Perceptions}
\format{A \code{\link{data.frame}}.}
\usage{
cola.perceptions
}
\description{
A modified version of \link{cola.perceptions.large}, where the data is in proportions and some columns have been removed.
}
\keyword{datasets}
