% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{colas}
\alias{colas}
\title{Colas}
\format{
A \code{\link{data.frame}}.
}
\usage{
colas
}
\description{
A data file from a survey of the Australian cola market in 2007. This is a modified version of \link{cola.original}.
}
\keyword{datasets}
