% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{csd.perceptions}
\alias{csd.perceptions}
\title{CSD Associations}
\format{
A \code{\link{data.frame}}.
}
\source{
Tim Bock.
}
\usage{
csd.perceptions
}
\description{
Binary associations between CSD brands and attributes, collected in Sydney c2000 amongst UTS students.
}
\keyword{datasets}
