% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{csd.situations}
\alias{csd.situations}
\title{CSD Situations}
\format{A \code{\link{data.frame}}.}
\source{
Tim Bock.
}
\usage{
csd.situations
}
\description{
Binary associations between CSD brands and the situations in which they are used, collected in Sydney c2001 amongst University of Sydney students.
}
\keyword{datasets}

