% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{hbat.with.splits}
\alias{hbat.with.splits}
\title{hbat.with.splits}
\format{A \code{\link{data.frame}}.}
\source{
Multivariate Data Analysis (7th Edition) 7th Edition, Joseph F. Hair Jr (Author), William C. Black  (Author), Barry J. Babin  (Author), Rolph E. Anderson (Author).
}
\usage{
hbat.with.splits
}
\description{
HBAT industries market researc].
}
\keyword{datasets}

