% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{hbatwithsplits}
\alias{hbatwithsplits}
\title{hbatwithsplits}
\format{A \code{\link{data.frame}}.}
\source{
Multivariate Data Analysis (7th Edition) 7th Edition, Joseph F. Hair Jr (Author), William C. Black  (Author), Barry J. Babin  (Author), Rolph E. Anderson (Author).
}
\usage{
hbatwithsplits
}
\description{
HBAT industries market research.
}
\keyword{datasets}
