% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{hotel.reviews}
\alias{hotel.reviews}
\title{Hotel Reviews}
\format{
A \code{\link{data.frame}} that contains two variables
\itemize{
    \item \code{Negative_Review} A persons negative comments about the hotel room/experience
    \item \code{Positive_Review} A persons positive comments about the hotel room/experience
}
}
\source{
22 August 2017: https://www.kaggle.com/jiashenliu/515k-hotel-reviews-data-in-europe Added to flipExampleData 28th August 2020
}
\usage{
hotel.reviews
}
\description{
The first 100,000 cases from the Hotels Review data scraped from booking.com by a user
   and uploaded to kaggle.com. It contains only two text variables from that dataset from a
    short form response to their hotel survey.
}
\keyword{datasets}
