% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ilock}
\alias{ilock}
\title{#' Q23 and Weights from Phone.sav
#'
#' 25 variables from a 5-point scale. Extra missing data has been added at random.
#' This makes up about 20% of the values. This is to test PCA and Factor analysis.
#'
#' @format A list containing:
#' \describe{
#'   \item{data.set}{25 variables from q23}
#'   \item{weight}{A vector of weights}
#' }
"pcaPhoneTestData"}
\format{
A \code{\link{data.frame}}.
}
\source{
Displayr Pty Ltd
Variables:
\itemize{
\item ResponseId label is Response ID
\item Status label is Status
\item Duration__in_seconds_ label is  Duration (in seconds)
\item UserLanguage label is   User Language
\item Panel label is  Panel
\item Q1 label is Gender
\item Q2 label is Age
\item Q3 label is State
\item Q4 label is Population density
\item Q5 label is Education
\item Q6a label is    White
\item Q6b label is    Hispanic
\item Q6c label is    Black
\item Q6d label is    Asian
\item oQ6 label is    Race Other specify
\item q6e label is    Race: Other
\item Q7 label is Family income
\item q9 label is Likes
\item q10 label is    Dislikes
\item q11 label is    Purchase intent
\item q12 label is    Uniqueness
\item q13 label is    Brand fit
\item q14 label is    Priced purchase intent
\item META_Browser label is   Browser Meta Info - Browser}
}
\usage{
ilock
}
\description{
iLock Concept Test Survey Data
}
\details{
#'  Default Q chart types.
#'
#' @format A \code{\link{vector}} containing RGB colors.
#' @source Numbers International Pty Ltd.
"q.colors"



 iLock
}
\keyword{datasets}
