% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{mobiles.feedback.4.weeks}
\alias{mobiles.feedback.4.weeks}
\title{US survey on mobile phones over a 4 week period concerning the
Likes and Dislikes of a Telecommunications company's services.}
\format{
data.frame
}
\usage{
mobiles.feedback.4.weeks
}
\description{
US survey on mobile phones over a 4 week period concerning the
Likes and Dislikes of a Telecommunications company's services.
}
\keyword{datasets}
