% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{mus03data}
\alias{mus03data}
\title{mus03data}
\format{An object of class \code{data.frame} with 3064 rows and 33 columns.}
\usage{
mus03data
}
\description{
@description mus03data dataset.
 @format A \code{\link{data.frame}}
}
\keyword{datasets}
