% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{park.visits}
\alias{park.visits}
\title{Park visits}
\format{
An object of class \code{data.frame} with 659 rows and 4 columns.
}
\usage{
park.visits
}
\description{
@description park visits dataset.
 @format A \code{\link{data.frame}}
}
\keyword{datasets}
