% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{pcaPhoneTestData}
\alias{pcaPhoneTestData}
\title{PCA Phone}
\format{An object of class \code{list} of length 4.}
\usage{
pcaPhoneTestData
}
\description{
@description PCA phone test dataset.
 @format A \code{\link{list}}
}
\keyword{datasets}
