% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{phone.associations}
\alias{phone.associations}
\title{Phone Associations}
\format{A \code{\link{data.frame}}.}
\source{
Tim Bock.

Adapted from Pinson, Christian, and Vikas Tibrewala (1996), “United Colors of Benetton,” INSEAD-CEDEP case study.
}
\usage{
phone.associations

phone.associations
}
\description{
Binary associations between Australian phone brands and attributes, collected from a sample of 725 friends and family of University of Sydney students, c. 2001. This is derived from \link{phone}.

Binary associations for Benetton.
}
\keyword{datasets}

