% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{phone}
\alias{phone}
\title{Phone}
\format{A \code{\link{data.frame}}.}
\source{
Tim Bock.
}
\usage{
phone
}
\description{
A survey about mobile phones, conducted from a sample of 725 friends and family of University of Sydney students, c. 2001.
}
\keyword{datasets}
