% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qColors}
\alias{qColors}
\title{Q's color palate (16 October 2014)}
\format{A vector containing RGB colors}
\usage{
qColors
}
\description{
Colors used in the data analysis app www.q-researchsoftware.com
}
\keyword{datasets}
