% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsvtable.R
\name{ReadCsvTable}
\alias{ReadCsvTable}
\title{Read CSV File as Table.
\code{ReadCsvTable} reads a CSV file and turns the first column into row names.table in where the first row and column contain labels.}
\usage{
ReadCsvTable(file, as.is = TRUE, ...)
}
\arguments{
\item{file}{A path and file name.}

\item{as.is}{Prevents factors from automatically being created (see \code{\link[utils]{read.csv}})}

\item{...}{Further arguments to be passed to read.csv}
}
\value{
A \code{\link[base]{data.frame}} containing a representation of the data in the file.
}
\description{
Read CSV File as Table.
\code{ReadCsvTable} reads a CSV file and turns the first column into row names.table in where the first row and column contain labels.
}
