% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{testweightsfilters}
\alias{testweightsfilters}
\title{testweightsfilters}
\format{A \code{\link{data.frame}}}
\usage{
testweightsfilters
}
\description{
testweightsfilter dataset.
}
\keyword{datasets}

