% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyspss.R
\name{TidySPSS}
\alias{TidySPSS}
\title{\code{TidySPSS}
Appends Variable Labels as attributes in a data.frame.}
\usage{
TidySPSS(data)
}
\arguments{
\item{data}{The \code{\link{data.frame}}, which contains the variable.labels and names attributes.}
}
\value{
A \code{\link{data.frame}}
}
\description{
\code{TidySPSS}
Appends Variable Labels as attributes in a data.frame.
}
