% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{weight}
\alias{weight}
\title{weight}
\format{A \code{\link{data.frame}}.}
\source{
6 August 2016: https://support.sas.com/documentation/cdl/en/statug/63033/HTML/default/viewer.htm#statug_introreg_sect003.htm.
}
\usage{
weight
}
\description{
Name, height, weight, and age, of children.
}
\keyword{datasets}
