% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cola.original}
\alias{cola.original}
\title{Cola Original}
\format{A \code{\link{data.frame}}.}
\usage{
cola.original
}
\description{
A data file from a survey of the Australian cola market in 2007.
}
\keyword{datasets}

